workaround vscode quirks
When starting a default client even if we do not pass a workspace folder to client options vs code will fall back to first workspace folder if any is open. This means that we cannot use default client for that case as it will result in 2 servers started for the same rootUri.
We need to pass a permissive document filter that will match untitled files to the first workspace client. Otherwise language server features won't work for such files. We cannot pass them to all clients as in that case in multiworkspace cases all servers would be handling the same messages.
On the other hand if no folder is open we have to use default client.