Add support for opening single files (#195)
* remove wrong comment

* extract client starting to function

make it possible to start default client for single file

* name started server for easier debugging

* do not start second client if vscode returns workspace folder for untitled file

this leads to broken behavior of 2 servers running with the same rootUri

* clear clients on deactivate to allow for garbage collection

* fix typespec

* make clienr finding behaviour consistent

* use consistent document selector on default and workspace clients

* rename

* workaround vscode quirks

When starting a default client even if we do not pass a workspace folder to client options vs code will fall back to first workspace folder if any is open. This means that we cannot use default client for that case as it will result in 2 servers started for the same rootUri.
We need to pass a permissive document filter that will match untitled files to the first workspace client. Otherwise language server features won't work for such files. We cannot pass them to all clients as in that case in multiworkspace cases all servers would be handling the same messages.
On the other hand if no folder is open we have to use default client.

* fix typo

* clarify comment

* clean up log messages

* add corresponding stop messages

* reintroduce fixed comment

* add comments