Put mix into quiet mode to surface git errors
I checked this on my machine and when I run git fetch without internet it exits with code 128 and prints to stderr
ssh: connect to host github.com port 22: Undefined error: 0
fatal: Could not read from remote repository.

Please make sure you have the correct access rights
and the repository exists.
However when this is run via Mix.SCM.Git.git! and we are running with Mix.Shell.IO we are getting an %IO.Stream{} here
https://github.com/elixir-lang/elixir/blob/6f96693b355a9b670f2630fd8e6217b69e325c5a/lib/mix/lib/mix/scm/git.ex#LL304C1-L304C1
so it’s getting printed to stderr. stderr is redirected to stdout for some reason in git! and hence we are redirecting stdout to /dev/null in elixir-ls launch scripts it’s getting swallowed and only the mix exception error is printed to stderr.

Solution: change mix shell and make the command return binary that gets printed to stderr

Fixes https://github.com/elixir-lsp/vscode-elixir-ls/issues/327